Faster fusion, better UI search
Before this, search would only search one name. However, due to the
entity fusion process, some people have multiple aliases within the
database. This patch creates a newline-separated `cached_names` field on
`FusedEntity` objects, and updates the UI to search this field. Now
entities like "scoder" + "stefan behlen" will both find the same fused
entity.

In the course of designing this improvement, a few other fusion issues
came up. First, the trigram index for fusion wasn't being used. Fixing
that cut about 80% off of the runtime of fusing people.

This patch also introduces commit-sha merges, where `git_commit`
entities in the database will be merged if the `commit_sha` field is
identical.

+++ lagoon/db/temptable.py
@@ -76,7 +76,7 @@ def _copy_table_from_model(model, multi_session):
 
 
 @contextlib.contextmanager
-def multi_session_dropper(sess_factory, table):
+def multi_session_dropper(sess_factory, table, *, keep=False):
     """Given some `table` from :meth:`temp_table_from_model` with
     ``multi_session=True``, produces a context manager that, on exit, uses
     `sess_factory` to destroy the table.
@@ -86,8 +86,9 @@ def multi_session_dropper(sess_factory, table):
     try:
         yield
     finally:
-        with sess_factory() as sess:
-            sess.execute(sa.text(f'''DROP TABLE "{table_name}"'''))
+        if not keep:
+            with sess_factory() as sess:
+                sess.execute(sa.text(f'''DROP TABLE "{table_name}"'''))
         _multi_session_tables.pop(table_name, None)
